Move distributed checkpointing to a subdirectory of experimental
