+++ torch_xla/experimental/distributed_checkpoint/planners.py
@@ -35,15 +35,10 @@ from torch.distributed.checkpoint.metadata import (
 from torch.distributed.checkpoint.utils import find_state_dict_object
 from torch.utils._pytree import tree_map
 from torch_xla.experimental.xla_sharding import XLAShardedTensor, XLAShard
-from torch_xla.experimental._distributed_checkpoint_helpers import (
+from torch_xla.experimental.distributed_checkpoint._helpers import (
     FLATTEN_MAPPING, flatten_state_dict, dedup_tensors, _is_sharded_tensor,
     set_element, narrow_tensor_by_index, _unwrap_xla_sharded_tensor, _CpuShards)
 from typing import Any, Dict, List, Tuple, Union
-
-__all__ = [
-    "SPMDSavePlanner",
-    "SPMDLoadPlanner",
-]
 
 
 class SPMDSavePlanner(SavePlanner):
